feat: Config screens for Forge & NeoForge.
