+++ src/main/java/mirsario/cameraoverhaul/entrypoints/NeoForgeInitializer.java
@@ -6,13 +6,18 @@
 /*package mirsario.cameraoverhaul.entrypoints;
 
 import mirsario.cameraoverhaul.*;
+import mirsario.cameraoverhaul.configuration.*;
+import net.neoforged.fml.*;
 import net.neoforged.fml.common.*;
+import net.neoforged.neoforge.client.gui.*;
 
 @Mod(CameraOverhaul.MOD_ID)
 @SuppressWarnings("unused")
 public class NeoForgeInitializer {
 	public NeoForgeInitializer() {
 		CameraOverhaul.onInitializeClient();
+
+        ModLoadingContext.get().registerExtensionPoint(IConfigScreenFactory.class, () -> (mc, p) -> ConfigScreen.getConfigScreen(p));
 	}
 }
 *///?}
