+++ versions/forge-1.17.1/gradle.properties
@@ -9,4 +9,6 @@ loader.id = forge
 loom.platform = forge
 
 # Dependencies
+mods.clothconfig.ref = 5.3.63
+mods.clothconfig.range = >=5.0.0
 deps.forge_loader = 1.17.1-37.1.1
