+++ versions/forge-1.19.0/gradle.properties
@@ -9,4 +9,6 @@ loader.id = forge
 loom.platform = forge
 
 # Dependencies
+mods.clothconfig.ref = 8.0.75
+mods.clothconfig.range = >=8.0.0
 deps.forge_loader = 1.19-41.1.0
