+++ versions/forge-1.19.3/gradle.properties
@@ -9,4 +9,6 @@ loader.id = forge
 loom.platform = forge
 
 # Dependencies
+mods.clothconfig.ref = 9.0.94
+mods.clothconfig.range = >=9.0.0
 deps.forge_loader = 1.19.3-44.1.0
