+++ versions/neoforge-1.20.6/gradle.properties
@@ -9,4 +9,6 @@ loader.id = neoforge
 loom.platform = neoforge
 
 # Dependencies
-deps.neoforge_loader=20.6.115
+mods.clothconfig.ref = 14.0.126
+mods.clothconfig.range = >=14.0.0
+deps.neoforge_loader = 20.6.115
