+++ versions/neoforge-1.21.2/gradle.properties
@@ -9,5 +9,7 @@ loader.id = neoforge
 loom.platform = neoforge
 
 # Dependencies
-deps.neoforge_loader=21.2.0-beta
+mods.clothconfig.ref = 16.0.141
+mods.clothconfig.range = >=16.0.0
+deps.neoforge_loader = 21.2.0-beta
